+++ internal/controli/tunnel_unix.go
@@ -52,7 +52,7 @@ func RunHostTunnelShellWithOptions(options TunnelHostOptions) int {
 		audit,
 		stats,
 		gate,
-		options.AuditInput,
+		options.HostOptions,
 		func(data []byte) error {
 			_, err := tty.Write(data)
 			return err